fix(widgets): break WidgetHost re-export cycle for chunk splitting
Rollup warning: `widgets/index.ts` re-exports `WidgetHost` from
`widgets/WidgetHost.tsx`, but the WidgetHost module also pulls other
widgets/* code that depends back through the barrel. When two consumers
end up in different chunks (e.g. the lazy AutomationsView chunk vs the
main shell), the cycle blocks proper splitting:

  Export "WidgetHost" of module ".../widgets/WidgetHost.tsx" was
  reexported through module ".../widgets/index.ts" while both modules
  are dependencies of each other and will end up in different chunks
  by current Rollup settings. This scenario is not well supported at
  the moment as it will produce a circular dependency between chunks
  and will likely lead to broken execution order.

Fix: switch the four non-widget callers (AutomationsView, HeartbeatsView,
CharacterHubView, TasksEventsPanel) from barrel imports to a direct
sub-path import of `widgets/WidgetHost`. The barrel itself stays for
non-cyclic exports (resolveWidgetsForSlot, declarations registry, etc.).

Build: still 0 warnings.
